ft_atoi.c updated

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-unsigned long	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	size_t i;
